refactoring routes with a controller
